refactor(auv-cli): remove inspect wrapper modules
